changed primary number type from double to float. This wil save space in hot paths and result in better locality

+++ RaychelEngine/include/Raychel/Core/RaychelMath/constants.h
@@ -30,25 +30,25 @@
 #pragma once
 
 namespace Raychel {
-	template<typename _float = double>
+	template<typename _float = float>
 	constexpr _float pi = _float(3.14159265358979323846264);
 
-	template<typename _float = double>
+	template<typename _float = float>
 	constexpr _float twoPi = 2 * pi<_float>;
 
-	template<typename _float = double>
+	template<typename _float = float>
 	constexpr _float halfPi = 0.5 * pi<_float>;
 
-	template<typename _float = double>
+	template<typename _float = float>
 	constexpr _float quarterPi = 0.25 * pi<_float>;
 
-	template<typename _float = double>
+	template<typename _float = float>
 	constexpr _float degToRad = pi<_float> / 180;
 
-	template<typename _float = double>
+	template<typename _float = float>
 	constexpr _float radToDeg = 180 / pi<_float>;
 
-	template<typename _float = double>
+	template<typename _float = float>
 	constexpr _float e = _float(2.718281828459045235360287);
 
 }
