+++ RaychelEngine/include/Raychel/Engine/Objects/lights.h
@@ -7,7 +7,7 @@ namespace Raychel {
 
     struct PointLight : public LightData
     {
-        PointLight(const Transform& _t, const color& _c, double _b, double _sz)
+        PointLight(const Transform& _t, const color& _c, float _b, float _sz)
             :LightData(_t, _c, _b, _sz)
         {}
