+++ RaychelEngine/src/Raychel/Engine/Objects/sdObjects.cpp
@@ -1,7 +1,7 @@
 #include "Raychel/Raychel.h"
 #include "Raychel/Engine/Objects/sdObjects.h"
 
-double Raychel::SdSphere::eval(const vec3& _p) const
+float Raychel::SdSphere::eval(const vec3& _p) const
 {
     return dist(_p, transform_.position) - radius;
 }
